requirements.txt
# Core Flask dependencies
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-SocketIO==5.3.6

# OAuth / OIDC
Authlib==1.3.1

# Database
SQLAlchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9

# Web server
gunicorn==23.0.0
eventlet==0.40.3

# Security and forms
Flask-WTF==1.2.1
Flask-Limiter==3.8.0

# Utilities
python-dotenv==1.0.0
pytz==2023.3
python-dateutil==2.8.2
Werkzeug==3.0.6
requests==2.32.4

# PDF Generation
WeasyPrint==60.2
Pillow==10.4.0
reportlab==4.0.7

# Background tasks
APScheduler==3.10.4

# Internationalization
Flask-Babel==4.0.0
Babel==2.14.0

# Development and testing
pytest==7.4.3
pytest-flask==1.3.0
black==24.8.0
flake8==6.1.0

# Security
cryptography==45.0.6
markdown==3.6
bleach==6.1.0
PyJWT==2.8.0
pyotp==2.9.0
qrcode==7.4.2

# Payment processing
stripe==7.9.0